[refactor] constant에 defaultOption추가
